% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getBioCyc.R
\name{getCycGenesList}
\alias{getCycGenesList}
\title{Get whole genes/proteins list}
\usage{
getCycGenesList(speID, type = "genes")
}
\arguments{
\item{speID}{The BioCyc species ID, for example "ECOLI" is for "Escherichia coli K-12 substr. MG1655".}

\item{type}{Get the "genes" or "proteins", and the default value is "genes".}
}
\value{
A vector of genes of proteins with BioCyc ID.
}
\description{
BioCyc Database API - Get the whole genes/proteins list from a given species
}
\details{
Get the BioCyc gene ID or protein ID list of a given species.
}
\examples{
getCycGenesList('ECOLI')
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

