% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBioCyc.R
\name{getCycPhylo}
\alias{getCycPhylo}
\title{Get species from BioCyc}
\usage{
getCycPhylo(speList, speType = "BioCyc", whole = FALSE, n = 2)
}
\arguments{
\item{speList}{A \code{character vector} contains species list like 'c("HUMAN", "ECOLI", "ZMOB579138")'. It should be consistent with the parameter `speType`.}

\item{speType}{A \code{string} "BioCyc" or "regexpr". "BioCyc" means it is the BioCyc species ID, for example "HUMAN" is the BioCyc ID for the Homo sapiens.The "regexpr" is used for regulare expression search with the Latin name for example "Escherichia coli".}

\item{whole}{A \code{logic} whether or not get the whole BioCyc species list, and the default value is FALSE.}

\item{n}{A \code{integer} refers to the number of cores.}
}
\value{
A \code{tbl_df} object contains BioCycID, Latin name, and version.
}
\description{
BioCyc Database API - Get species information from BioCyc.
}
\details{
Get the BioCyc species information including the BioCyc species ID and the Latin name.
}
\examples{
## search species list from BioCyc ID
getCycPhylo(c('HUMAN', 'ECOLI', 'ZMOB579138'), speType = 'BioCyc')

## search species whose names include 'Escherichia coli'
getCycPhylo('Escherichia coli', speType = 'regexpr')

\dontrun{
## get whole BioCyc species information table
getCycPhylo(whole = TRUE)}
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
