% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getBioCyc.R
\name{getCycSpeTU}
\alias{getCycSpeTU}
\title{Get whole TU information from a given species}
\usage{
getCycSpeTU(speID)
}
\arguments{
\item{speID}{The BioCyc species ID, for example "ECOLI" is for "Escherichia coli K-12 substr. MG1655".}
}
\value{
The TU id vector
}
\description{
BioCyc Database API - Get whole TU list of a given species from BioCyc database.
Get the whole transcription units from a given species. It may take more than 10 minutes to retrieve the xml file.
}
\examples{
\dontrun{
## get Streptococcus mutans UA159 TU
SMTU <- getCycSpeTU('SMUT210007')
}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

