% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TU.R
\name{getCycTU}
\alias{getCycTU}
\title{TU}
\usage{
getCycTU(speID)
}
\arguments{
\item{speID}{The BioCyc species ID, for example "ECOLI" is for "Escherichia coli K-12 substr. MG1655".}
}
\value{
A vector of TU ids.
}
\description{
BioCyc Database API - Get whole transcription unit list of a given species from BioCyc database.
}
\details{
Get transcription units from a given species. It may take more than 10 minutes to retrieve the xml file.
}
\examples{
## get Streptococcus mutans UA159 TU
smTU <- getCycTU('SMUT210007')
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
