% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TU.R
\name{getCycTUfGene}
\alias{getCycTUfGene}
\title{Get TU from ID.}
\usage{
getCycTUfGene(geneID, evidence = FALSE)
}
\arguments{
\item{geneID}{A BioCyc gene ID with the length of 1.}

\item{evidence}{Logical value indicates whether to return the evidence value.}
}
\value{
A \code{list} contains TUs or NULL
}
\description{
BioCyc Database API - Get transcription unit (TU) from gene
}
\details{
Get TU from a given BioCyc gene ID.
If the given gene has no TU, "NULL" will be returned. If the "evidence" is set to TRUE, a list will return.
}
\examples{
getCycTUfGene('ECOLI:EG10102')
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
