% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TU.R
\name{getCycTUInfo}
\alias{getCycTUInfo}
\title{Get one TU information}
\usage{
getCycTUInfo(TUID)
}
\arguments{
\item{TUID}{A BioCyc TU ID with the length of 1.}
}
\value{
A list of TU information.
}
\description{
BioCyc Database API - Get TU information from BioCyc database.
}
\details{
Get TU information including genes in TU and evidence.
There is another way to get the genes "http://biocyc.org/apixml?fn=transcription-unit-genes&id=ECOLI:TU0-42328&detail=full".
}
\examples{
getCycTUInfo('ECOLI:TU0-6636')
getCycTUInfo('ECOLI:TU00260')
getCycTUInfo('SMUT210007:TU1FZX-806')
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
