% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBioCyc.R
\name{KEGG2Sym}
\alias{KEGG2Sym}
\title{KEGG gene ID to symbol.}
\usage{
KEGG2Sym(geneKEGG)
}
\arguments{
\item{geneKEGG}{The gene symbol extract from KEGG.}
}
\value{
A \code{character vector} including the raw KEGGID.
}
\description{
KEGG gene ID to symbol
}
\details{
Try to transfer KEGG ID to symbol, if it has one; otherwise, we just use the `KEGGID`. The return gene information is split by ";" and ",". All elements, expect ones contain space, are returned.
}
\examples{
KEGG2Sym('smu:SMU_23')

KEGG2Sym('smu:SMU_t02')

KEGG2Sym('smu:SMU_24')

KEGG2Sym('smu:SMU_20')

KEGG2Sym('smu:SMU_408')
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
