% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{getCycPathway}
\alias{getCycPathway}
\alias{getCycGenesfPathway}
\title{Pathway}
\usage{
getCycPathway(speID)

getCycGenesfPathway(pathID)
}
\arguments{
\item{speID}{The BioCyc species ID, for example "ECOLI" is for "Escherichia coli K-12 substr. MG1655".}

\item{pathID}{Pathway ID.}
}
\value{
\itemize{
  \item \code{getCycPathway()}:  A \code{character vector} indicates pathway ids.
  \item \code{getCycGenesfPathway()}: A \code{list} indicates genes, proteins, and common names.
}
}
\description{
BioCyc Database API - Pathway
}
\details{
\itemize{
  \item \code{getCycPathway()}: Get whole pathways list.
  \item \code{getCycGenesfPathway()}: Get genes from a given pathway ID.
}

It may take more than 10 minutes to retrieve the xml file.
}
\examples{
## Candida albicans SC5314 pathways
calTU <- getCycPathway('CALBI')

## genes of pathway "CALBI:PWY3B3-8"
genes <- getCycGenesfPathway('CALBI:PWY3B3-8')
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
