% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBioCycAdd.R
\name{PhyloCyc2NCBI}
\alias{PhyloCyc2NCBI}
\title{Get NCBI Taxonomy ID From BioCyc species ID}
\usage{
PhyloCyc2NCBI(speID)
}
\arguments{
\item{speID}{The BioCyc species ID, for example "ECOLI" is for "Escherichia coli K-12 substr. MG1655".}
}
\value{
The corresponding NCBI Taxonomy ID.
}
\description{
BioCyc Database Additional API - Get the NCBI taxonomy ID from a given BioCyc ID
}
\details{
NCBI taxonomy ID is used as unique ID accoss cyc and BioCyc databases. This functions is used to get the corresponding NCBI Taxonomy ID from BioCyc. If BioCyc has no official NCBI taxonomy ID, it will return a character with length of 0 (it looks like "").
}
\examples{
## get human, mouse, and Ecoli NCBI taxonomy ID.
PhyloCyc2NCBI('HUMAN')
PhyloCyc2NCBI('MOUSE')
PhyloCyc2NCBI('ECOLI')

}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
