% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{testLen}
\alias{testLen}
\title{Test length is 0 or not}
\usage{
testLen(inputVal, trueVal, falseVal)
}
\arguments{
\item{inputVal}{vector}

\item{trueVal}{return this value, if the length of "inputVal" is 0.}

\item{falseVal}{return this value, if the length of "falseVal" is not 0.}
}
\value{
"trueVal" or "falseVal"
}
\description{
Test if the input vector's length is 0
}
\details{
To test the length of input vector, if the length is 0, return "trueVal", else return "falseval"
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}

