% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getBioCyc.R
\name{transGeneIDKEGG2Cyc}
\alias{transGeneIDKEGG2Cyc}
\title{Transfer KEGG ID to BioCyc ID.}
\usage{
transGeneIDKEGG2Cyc(KEGGID, speKEGGID, speCycID, type = "gene")
}
\arguments{
\item{KEGGID}{Only one KEGG ID}

\item{speKEGGID}{Species BioCyc ID.}

\item{speCycID}{Species KEGG ID}

\item{type}{'gene' or 'protein'}
}
\value{
The BioCyc gene ID or "0", if gene is not found.
}
\description{
Translate KEGG ID to BioCyc ID.
}
\details{
Translate the KEGG gene ID to BioCyc gene ID is tricky. In BioCyc, the gene names is not in a uniform; some of them use symbol like "dnaK", but some use the KEGGID. For genes, if symbol is given, we use 'http://biocyc.org/xmlquery?query=[x:x<-ECOLI^^genes,x^name="atpA"]&detail=full'. There are two circumstances that will return "0": one is that  BioCyc database may marker some genes as "Pseudo-Genes", and the other is different gene symbols in KEGG and BioCyc. For proteins, we at first transfer KEGG gene IDs to UniProt IDs, and then to BioCyc gene IDs.
}
\examples{
## symbol is "atpD"
transGeneIDKEGG2Cyc('b3732', 'eco', 'ECOLI')

## symbol is "SMU_408" but the first annotation word is "permease"
transGeneIDKEGG2Cyc('SMU_408', 'smu', 'SMUT210007')

## It will return "0" because of the symbol 'atpE_H' from KEGG.
## The symbol in BioCyc is 'atpE/H'.
transGeneIDKEGG2Cyc('Bd0010', 'bba', 'BBAC264462')

## retrieve protein
transGeneIDKEGG2Cyc('b0001', 'eco', 'ECOLI', type = 'protein')
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

