% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBioCycAdd.R
\name{transPhyloCyc2NCBI}
\alias{transPhyloCyc2NCBI}
\title{Get NCBI Taxonomy ID From BioCyc species ID}
\usage{
transPhyloCyc2NCBI(cycID, n = 4)
}
\arguments{
\item{cycID}{The cyc species ID. The KEGG support multiple species ID, for example c('HUMAN', 'MOUSE', 'ECOLI').}

\item{n}{The number of CPUs or processors, and the default value is 4.}
}
\value{
The corresponding NCBI Taxonomy ID in character vector.
}
\description{
BioCyc Database Additional API - Get the NCBI taxonomy ID from a given BioCyc ID
}
\details{
NCBI taxonomy ID is used as unique ID accoss cyc and BioCyc databases. This functions is used to get the corresponding NCBI Taxonomy ID from BioCyc. If BioCyc has no official NCBI taxonomy ID, it will return a character with length of 0 (it looks like "").
}
\examples{
## get human, mouse, and Ecoli NCBI taxonomy ID with 2 threads
transPhyloCyc2NCBI(c('HUMAN', 'MOUSE', 'ECOLI'), n = 2)

\dontrun{
## transfer all BioCyc species ID to NCBI taxonomy ID
wBiocycSpe <- getCycPhylo(whole = TRUE)
wNCBISpe <- transPhyloCyc2NCBI(wBiocycSpe[, 1])
}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

