% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{xmlNodeAttr}
\alias{xmlNodeAttr}
\title{Get all nodes attributes directly from the XML file.}
\usage{
xmlNodeAttr(xmlFile, nodePath, attrName)
}
\arguments{
\item{xmlFile}{XML file.}

\item{nodePath}{The XPath of nodeset (one or mutiple nodes).}

\item{attrName}{Attributes name}
}
\value{
Nodeset attributes
}
\description{
Get attribute values.
}
\details{
Get attribute values from the XML file.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}

