% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{xmlNodeVal}
\alias{xmlNodeVal}
\title{Get all nodes value directly from the XML file.}
\usage{
xmlNodeVal(xmlFile, nodePath)
}
\arguments{
\item{xmlFile}{XML file.}

\item{nodePath}{The XPath of nodeset (one or mutiple nodes).}
}
\value{
Nodeset value
}
\description{
Get node values.
}
\details{
Get node values from the XML file.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}

