% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBioCyc.R
\name{getCycGeneInfo}
\alias{getCycGeneInfo}
\title{Get one gene information from BioCyc.}
\usage{
getCycGeneInfo(geneID)
}
\arguments{
\item{geneID}{A BioCyc gene ID with the length of 1.}
}
\value{
A list.
}
\description{
BioCyc Database API - Get gene information from BioCyc database.
}
\details{
The gene information from BioCyc including genome location and gene name information. Some genes in BioCyc may do not have common names or accession names In this circumstance, "NULL" will be return
}
\examples{
## get "atpE" gene information from Ecoli K-12 MG1655 strain.
getCycGeneInfo('ECOLI:EG10102')
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
