% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getBioCyc.R
\name{getCycTUfGene}
\alias{getCycTUfGene}
\title{Get TU from ID.}
\usage{
getCycTUfGene(geneID, speID, evidence = FALSE)
}
\arguments{
\item{geneID}{A BioCyc gene.}

\item{speID}{The BioCyc species ID, for example "ECOLI" is for "Escherichia coli K-12 substr. MG1655".}

\item{evidence}{Logical value indicates whether to return the evidence value.}
}
\value{
A vector contains TUs or NULL
}
\description{
BioCyc Database API - Get transcription unit (TU) from gene
}
\details{
Get TU from a given BioCyc gene ID.
If the given gene has no TU, "NULL" will be returned. If the "evidence" is set to TRUE, a list will return.
}
\examples{
getCycTUfGene('EG10102', 'ECOLI')
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

