% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TU.R
\name{getCycTUInfo}
\alias{getCycTUInfo}
\title{Get one TU information}
\usage{
getCycTUInfo(TUID, speID)
}
\arguments{
\item{TUID}{A BioCyc TU ID with the length of 1.}

\item{speID}{The BioCyc species ID, for example "ECOLI" is for "Escherichia coli K-12 substr. MG1655".}
}
\value{
A list of TU information.
}
\description{
BioCyc Database API - Get TU information from BioCyc database.
}
\details{
Get TU information including genes in TU and evidence.
There is another way to get the genes "http://biocyc.org/apixml?fn=transcription-unit-genes&id=ECOLI:TU0-42328&detail=full".
}
\examples{
getCycTUInfo('TU0-6636', 'ECOLI')
getCycTUInfo('TU00260', 'ECOLI')
getCycTUInfo('TUC7Z-43', 'SMUT210007')
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

