% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBioCyc.R
\name{Sym2BioCyc}
\alias{Sym2BioCyc}
\title{KEGG symbol to BioCyc ID.}
\usage{
Sym2BioCyc(symbol, speID)
}
\arguments{
\item{symbol}{The gene symbol extract from KEGG.}

\item{speID}{The BioCyc species ID, for example "ECOLI" is for "Escherichia coli K-12 substr. MG1655".}
}
\value{
A \code{list}. The BioCyc gene ID or "0", if gene is not found.
}
\description{
KEGG symbol to Biocyc gene ID
}
\details{
Try to convert the KEGG symbol to BioCyc gene ID.
}
\examples{
Sym2BioCyc('dnaA', 'SMUT210007')

Sym2BioCyc('SMU_06', 'SMUT210007')
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
